requirements.txt

pillow==10.3.0
qrcode==7.4.2
